V1.0.0 Official
Added all unity.mathematics noise functions with complete control over
the color and extra input parameters.

added ValueInterpertation which allows the user to control which return
values of the return function go to wo which rgb value of the texture.
(i like  setting y and z to true and x to false for those cyan values)

fixed some missing functions and incorrect values.

updated the Editor to somewhat seemlessly transition from
noise to noise.
Note: for width,height,length,and depth, they are all for loops so increasing
	them above 64 (with length and depth in mind) will cause the calculations to take longer
	to computer. you can reduce this by changing the Min value for these values.